moving  uploads to properties
