+++ api/server.js
@@ -64,43 +64,43 @@ server.post('/single', upload.single('test'), (req, res) => {
 
 /** AWS catalog */
 
-aws.config.update({
-  secretAccessKey: process.env.AWS_SECRET_ACCESS_KEY,
-  accessKeyId: process.env.AWS_ACCESS_KEY_ID,
-  region: 'us-west-1'
-});
+// aws.config.update({
+//   secretAccessKey: process.env.AWS_SECRET_ACCESS_KEY,
+//   accessKeyId: process.env.AWS_ACCESS_KEY_ID,
+//   region: 'us-west-1'
+// });
 
-const s3 = new aws.S3();
-const awsStorage = multerS3({
-  s3: s3,
-  bucket: process.env.AWS_BUCKET_NAME,
-  key: function(req, file, cb) {
-    console.log(file);
-    cb(null, file.originalname);
-  }
-});
+// const s3 = new aws.S3();
+// const awsStorage = multerS3({
+//   s3: s3,
+//   bucket: process.env.AWS_BUCKET_NAME,
+//   key: function(req, file, cb) {
+//     console.log(file);
+//     cb(null, file.originalname);
+//   }
+// });
 
-const upload = multer({
-  /**if you are using local storage than use
-   * storage: fileStorage,
-   * if you are using aws storage than use
-   * storage: awsStorage,
-   */
-  storage: awsStorage,
-  limits: { fileSize: 5000000 }
-  // fileFilter: function(req, file, cb) {
-  //   checkFileType(file, cb);
-  // }
-});
-server.post('/upload/:id', upload.single('profile'), (req, res, err) => {
-  try {
-    const imageUrl = req.file.location;
-    const property_id = req.params.id;
-    const package = { property_id, imageUrl };
-    res.send(package);
-  } catch (err) {
-    res.send(400);
-  }
-});
+// const upload = multer({
+//   /**if you are using local storage than use
+//    * storage: fileStorage,
+//    * if you are using aws storage than use
+//    * storage: awsStorage,
+//    */
+//   storage: awsStorage,
+//   limits: { fileSize: 5000000 }
+//   // fileFilter: function(req, file, cb) {
+//   //   checkFileType(file, cb);
+//   // }
+// });
+// server.post('/upload/:id', upload.single('profile'), (req, res, err) => {
+//   try {
+//     const imageUrl = req.file.location;
+//     const property_id = req.params.id;
+//     const package = { property_id, imageUrl };
+//     res.send(package);
+//   } catch (err) {
+//     res.send(400);
+//   }
+// });
 
 module.exports = server;
